% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttongrass.R
\name{buttongrass_fire}
\alias{buttongrass_fire}
\title{Buttongrass Fire Behaviour (Marsden-Smedley 2003)}
\usage{
buttongrass_fire(
  temp,
  rh,
  wind_speed,
  rain_amount = 0,
  hours_since_rain = 0,
  age = 20,
  cover = 0,
  productivity = 2,
  slope = 0
)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}

\item{wind_speed}{Wind speed at 10 m (km/h).}

\item{rain_amount}{Amount of recent rain (mm, default 0).}

\item{hours_since_rain}{Hours since rain stopped (default 0).}

\item{age}{Buttongrass age (years, default 20).}

\item{cover}{Fuel cover (\%, 0 if unknown).}

\item{productivity}{Site productivity (1 = low, 2 = medium, default 2).}

\item{slope}{Slope (degrees, default 0).}
}
\value{
A list with fire behaviour outputs.
}
\description{
Buttongrass Fire Behaviour (Marsden-Smedley 2003)
}
\references{
Marsden-Smedley, J.B. (2003). Buttongrass moorland fire
behaviour model.
}
