% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttongrass.R
\name{buttongrass_fuel_load}
\alias{buttongrass_fuel_load}
\title{Buttongrass Fuel Load}
\usage{
buttongrass_fuel_load(age = 20, cover = 0, productivity = 2)
}
\arguments{
\item{age}{Buttongrass age since last fire (years, default 20).}

\item{cover}{Foliage projective cover (\%, 0 if unknown).}

\item{productivity}{Site productivity code (1 = low, 2 = medium).}
}
\value{
Fuel load (t/ha).
}
\description{
Buttongrass Fuel Load
}
