% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppression.R
\name{combined_suppression_rate}
\alias{combined_suppression_rate}
\title{Combined suppression rate}
\usage{
combined_suppression_rate(
  hand = 0,
  d4 = 0,
  d6 = 0,
  tanker = 0,
  bomber1 = 0,
  bomber2 = 0
)
}
\arguments{
\item{hand}{Hand trail rate (m/h, default 0).}

\item{d4}{D4 dozer rate (m/h, default 0).}

\item{d6}{D6 dozer rate (m/h, default 0).}

\item{tanker}{Tanker rate (m/h, default 0).}

\item{bomber1}{Firebomber rate 1 (m/h, default 0).}

\item{bomber2}{Firebomber rate 2 (m/h, default 0).}
}
\value{
Total suppression rate (m/h).
}
\description{
Combined suppression rate
}
