% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppression.R
\name{d6_dozer_rate}
\alias{d6_dozer_rate}
\title{D6 bulldozer construction rate}
\usage{
d6_dozer_rate(n_dozers, slope, debris, flame_height)
}
\arguments{
\item{n_dozers}{Number of D6 bulldozers.}

\item{slope}{Slope (degrees).}

\item{debris}{Surface debris level: "None", "Some", "Significant",
or "Very Significant".}

\item{flame_height}{Flame height (m).}
}
\value{
Estimated construction rate (m/h).
}
\description{
D6 bulldozer construction rate
}
