% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{dew_point}
\alias{dew_point}
\title{Calculate dew point from temperature and relative humidity}
\usage{
dew_point(temp, rh)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}
}
\value{
Dew point temperature (degrees C).
}
\description{
Calculate dew point from temperature and relative humidity
}
\references{
Magnus formula with coefficients a = 17.2694, b = 237.7.
}
