% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought.R
\name{drought_factor}
\alias{drought_factor}
\title{Calculate drought factor}
\usage{
drought_factor(kbdi, days_since_rain, rain_amount)
}
\arguments{
\item{kbdi}{Keetch-Byram Drought Index or Soil Dryness Index.}

\item{days_since_rain}{Number of days since last rain.}

\item{rain_amount}{Amount of rain (mm).}
}
\value{
Drought factor (0-10).
}
\description{
Drought factor based on KBDI/SDI, days since rain, and rainfall amount.
Capped at 10.
}
