% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{fire_spread_direction}
\alias{fire_spread_direction}
\title{Calculate the general direction of fire spread}
\usage{
fire_spread_direction(wind_dir_deg)
}
\arguments{
\item{wind_dir_deg}{Wind direction in degrees from true north.}
}
\value{
Direction of fire spread in degrees.
}
\description{
Fire spreads in the opposite direction to the wind.
}
