% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppression.R
\name{firebomber_rate}
\alias{firebomber_rate}
\title{Firebomber holding rate}
\usage{
firebomber_rate(n_bombers, turnaround, aircraft_type)
}
\arguments{
\item{n_bombers}{Number of aircraft.}

\item{turnaround}{Turnaround time (minutes).}

\item{aircraft_type}{Aircraft type: "1400L Medium Helicopter" or
"2500L Fixed-wing".}
}
\value{
Estimated holding rate (m/h).
}
\description{
Firebomber holding rate
}
