% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_mcarthur.R
\name{forest_leaflet80}
\alias{forest_leaflet80}
\title{McArthur Leaflet 80 Control Burning Guide}
\usage{
forest_leaflet80(
  temp,
  rh,
  wind_speed,
  drought_factor,
  fuel_load,
  slope = 0,
  wind_reduction = 3,
  time_hour = 14,
  kbdi = 50,
  days_since_rain = 5,
  rain_amount = 0,
  use_bom_df = FALSE
)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}

\item{wind_speed}{Wind speed at 10 m in the open (km/h).}

\item{drought_factor}{Drought factor (0-10).}

\item{fuel_load}{Fine fuel load (t/ha).}

\item{slope}{Slope (degrees, default 0).}

\item{wind_reduction}{Wind reduction factor (1-6, default 3).}

\item{time_hour}{Hour of day (0-23). Used to select FMC equation.}

\item{kbdi}{Keetch-Byram Drought Index.}

\item{days_since_rain}{Days since last rain.}

\item{rain_amount}{Amount of last rain (mm).}

\item{use_bom_df}{Logical. If TRUE, use the provided drought factor
directly for fuel availability.}
}
\value{
A list with fire behaviour outputs.
}
\description{
Rate of spread and related outputs under low-intensity conditions.
Uses equations published by Gould (1994).
}
\references{
Gould, J.S. (1994). Evaluation of McArthur's control burning
guide in regrowth \emph{Eucalyptus sieberi} forest. \emph{Aust. Forestry}
57, 86-93.
}
