% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_mcarthur.R
\name{forest_mcarthur_mk5}
\alias{forest_mcarthur_mk5}
\title{McArthur Mk5 Forest Fire Behaviour}
\usage{
forest_mcarthur_mk5(
  temp,
  rh,
  wind_speed,
  drought_factor,
  fuel_load,
  slope = 0,
  wind_reduction = 3
)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}

\item{wind_speed}{Wind speed at 10 m in the open (km/h).}

\item{drought_factor}{Drought factor (0-10).}

\item{fuel_load}{Fine fuel load (t/ha).}

\item{slope}{Slope (degrees, default 0).}

\item{wind_reduction}{Wind reduction factor (1-6, default 3).}
}
\value{
A list with fire behaviour outputs.
}
\description{
Calculates rate of spread, flame height, spotting distance, fireline
intensity and heat output based on McArthur Mk5.
}
