% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_vesta.R
\name{forest_vesta}
\alias{forest_vesta}
\title{VESTA Forest Fire Behaviour}
\usage{
forest_vesta(
  temp,
  rh,
  wind_speed,
  slope,
  surface_score,
  near_surface_score,
  near_surface_height,
  elevated_score,
  elevated_height,
  bark_score,
  month = 1,
  hour = 14
)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}

\item{wind_speed}{Wind speed at 10 m in the open (km/h).}

\item{slope}{Slope (degrees).}

\item{surface_score}{Surface fuel hazard score (0-4).}

\item{near_surface_score}{Near-surface fuel hazard score (0-4).}

\item{near_surface_height}{Near-surface fuel height (cm).}

\item{elevated_score}{Elevated fuel hazard score (0-4).}

\item{elevated_height}{Elevated fuel height (m).}

\item{bark_score}{Bark fuel hazard score (0-4).}

\item{month}{Month (1-12, default 1).}

\item{hour}{Hour of day (0-23, default 14).}
}
\value{
A list with fire behaviour outputs.
}
\description{
Rate of spread model for dry eucalypt forest.
}
\references{
Cheney et al. (2007) VESTA fire behaviour model.
}
