% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{fuel_moisture_mcarthur}
\alias{fuel_moisture_mcarthur}
\title{Estimate surface fine fuel moisture content (McArthur)}
\usage{
fuel_moisture_mcarthur(temp, rh)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}
}
\value{
Fine fuel moisture content (\%).
}
\description{
Estimate surface fine fuel moisture content (McArthur)
}
