% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.R
\name{grass_fdi}
\alias{grass_fdi}
\title{CSIRO Grassland Fire Danger Index}
\usage{
grass_fdi(curing, temp, wind_speed, rh)
}
\arguments{
\item{curing}{Degree of grass curing (\%).}

\item{temp}{Air temperature (degrees C).}

\item{wind_speed}{Wind speed at 10 m (km/h).}

\item{rh}{Relative humidity (\%).}
}
\value{
Grassland FDI.
}
\description{
CSIRO Grassland Fire Danger Index
}
\references{
Purton (1982) BoM Meteorological Note No.147.
}
