% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.R
\name{grass_fire}
\alias{grass_fire}
\title{CSIRO Grassland Fire Behaviour}
\usage{
grass_fire(temp, rh, wind_speed, curing, slope = 0)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}

\item{wind_speed}{Wind speed at 10 m (km/h).}

\item{curing}{Degree of grass curing (\%).}

\item{slope}{Slope (degrees, default 0).}
}
\value{
A list with fire behaviour outputs for all five grass types.
}
\description{
Implements the CSIRO grassland fire spread model with Cruz et al. (2015)
curing function. Five sub-models: natural, grazed, eaten-out, open
woodland, and woodland.
}
\references{
Cheney, Gould & Catchpole (1998). Prediction of fire spread
in grasslands. \emph{Int. J. Wildland Fire} 8(1):1-13.

Cruz et al. (2015). Updated curing function.
}
