% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppression.R
\name{hand_trail_rate}
\alias{hand_trail_rate}
\title{Hand trail construction rate}
\usage{
hand_trail_rate(n_firefighters, slope, elevated_fuel, flame_height)
}
\arguments{
\item{n_firefighters}{Number of fire fighters.}

\item{slope}{Slope (degrees).}

\item{elevated_fuel}{Elevated fuel hazard: "Low", "Moderate", "High",
or "Very High".}

\item{flame_height}{Flame height (m).}
}
\value{
Estimated construction rate (m/h).
}
\description{
Estimated fireline construction rate by hand crews over a 12-hour shift.
}
\references{
McCarthy, Tolhurst & Wouters (2003). Research Report No.56,
Fire Management, DSE Victoria.
}
