% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heath.R
\name{heath_fire}
\alias{heath_fire}
\title{Heathland Fire Behaviour (Anderson et al. 2015)}
\usage{
heath_fire(
  temp,
  rh,
  wind_speed,
  veg_height,
  slope = 0,
  overcast = FALSE,
  hour = 14,
  month = 2
)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}

\item{wind_speed}{Wind speed at 10 m in the open (km/h).}

\item{veg_height}{Vegetation top height (m).}

\item{slope}{Slope (degrees, default 0).}

\item{overcast}{Logical. If TRUE, delta correction for overcast sky is
disabled (default FALSE).}

\item{hour}{Hour of day (0-23, default 14). Used for delta correction.}

\item{month}{Month (1-12, default 2). Used for delta correction.}
}
\value{
A list with heath and woodland fire behaviour outputs.
}
\description{
Heathland Fire Behaviour (Anderson et al. 2015)
}
\references{
Anderson et al. (2015). A generic, empirical-based model for
predicting rate of fire spread in shrublands. \emph{Int. J. Wildland Fire}.

Catchpole et al. (1998). Cooperative development of equations for
heathland fire behaviour.
}
