% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallee.R
\name{mallee_fire}
\alias{mallee_fire}
\title{WA Mallee Fire Spread (McCaw 1998)}
\usage{
mallee_fire(wind_speed, dead_fmc, slope = 0)
}
\arguments{
\item{wind_speed}{Wind speed at 10 m (km/h).}

\item{dead_fmc}{Dead fuel moisture content (\%).}

\item{slope}{Slope (degrees, default 0).}
}
\value{
A list with rate of spread (m/h).
}
\description{
Model developed for mature (24-year-old) mallee with relatively high
fuel continuity.
}
\references{
McCaw, W.L. (1998). Fire spread in WA mallee.
}
