% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallee.R
\name{mallee_heath_fire}
\alias{mallee_heath_fire}
\title{Mallee-Heath Fire Behaviour (Cruz et al. 2013)}
\usage{
mallee_heath_fire(
  temp,
  rh,
  wind_speed,
  wind_speed_2m = NULL,
  surface_fmc,
  overstorey_cover,
  overstorey_height,
  surface_fuel_load,
  slope = 0
)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}

\item{wind_speed}{Wind speed at 10 m (km/h).}

\item{wind_speed_2m}{Wind speed at 2 m (km/h, optional). If NULL, only
the 10 m wind model is used.}

\item{surface_fmc}{Surface fine fuel moisture content (\%).}

\item{overstorey_cover}{Overstorey cover (\%).}

\item{overstorey_height}{Overstorey height (m).}

\item{surface_fuel_load}{Combined surface fine fuel load (t/ha).}

\item{slope}{Slope (degrees, default 0).}
}
\value{
A list with fire behaviour outputs.
}
\description{
Combines WA and SA experimental data for mallee-heath shrublands.
}
\references{
Cruz, M.G., McCaw, W.L. et al. (2013). \emph{Environmental
Modelling & Software} 40:21-34.
}
