% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_mcarthur.R
\name{mcarthur_fdi}
\alias{mcarthur_fdi}
\title{McArthur Mk5 Forest Fire Danger Index}
\usage{
mcarthur_fdi(drought_factor, rh, temp, wind_speed, wind_reduction = 3)
}
\arguments{
\item{drought_factor}{Drought factor (0-10).}

\item{rh}{Relative humidity (\%).}

\item{temp}{Air temperature (degrees C).}

\item{wind_speed}{Wind speed at 10 m in the open (km/h).}

\item{wind_reduction}{Wind reduction factor from 10 m open to 2 m in
forest (1-6, default 3).}
}
\value{
Forest Fire Danger Index (FDI).
}
\description{
McArthur Mk5 Forest Fire Danger Index
}
\references{
Noble, Bary & Gill (1980). McArthur's fire danger meters
expressed as equations. \emph{Aust. J. Ecol.} 5:201-203.
}
