% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_mcarthur.R
\name{mcarthur_fdi_matthews}
\alias{mcarthur_fdi_matthews}
\title{McArthur FDI using Matthews fuel moisture method}
\usage{
mcarthur_fdi_matthews(drought_factor, wind_speed, ffmc)
}
\arguments{
\item{drought_factor}{Drought factor (0-10).}

\item{wind_speed}{Wind speed at 10 m in the open (km/h).}

\item{ffmc}{Fine fuel moisture content (\%).}
}
\value{
FDI value.
}
\description{
McArthur FDI using Matthews fuel moisture method
}
