% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook.R
\name{redbook_aff_karri}
\alias{redbook_aff_karri}
\title{WA Red Book Available Fuel Factor (Karri)}
\usage{
redbook_aff_karri(smc, profile_mc)
}
\arguments{
\item{smc}{Surface moisture content (\%).}

\item{profile_mc}{Profile moisture content (\%).}
}
\value{
Available fuel factor (0-1).
}
\description{
WA Red Book Available Fuel Factor (Karri)
}
