% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook.R
\name{redbook_bdu}
\alias{redbook_bdu}
\title{WA Red Book Basic Drying Unit}
\usage{
redbook_bdu(temp, rh)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}
}
\value{
Basic drying unit value.
}
\description{
WA Red Book Basic Drying Unit
}
