% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook.R
\name{redbook_ddc_jarrah_surface}
\alias{redbook_ddc_jarrah_surface}
\title{WA Red Book Day Drying Correction (Jarrah Surface)}
\usage{
redbook_ddc_jarrah_surface(smc_0800, bdu)
}
\arguments{
\item{smc_0800}{0800h surface moisture content (\%).}

\item{bdu}{Basic drying unit.}
}
\value{
Day drying correction.
}
\description{
WA Red Book Day Drying Correction (Jarrah Surface)
}
