% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook.R
\name{redbook_ddc_karri_profile}
\alias{redbook_ddc_karri_profile}
\title{WA Red Book Day Drying Correction (Karri Profile)}
\usage{
redbook_ddc_karri_profile(pmc_0800, bdu)
}
\arguments{
\item{pmc_0800}{0800h profile moisture content (\%).}

\item{bdu}{Basic drying unit.}
}
\value{
Day drying correction for Karri profile.
}
\description{
WA Red Book Day Drying Correction (Karri Profile)
}
