% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook.R
\name{redbook_fdi_jarrah}
\alias{redbook_fdi_jarrah}
\title{WA Red Book Northern Jarrah FDI}
\usage{
redbook_fdi_jarrah(smc, wind_speed, wind_reduction = 3)
}
\arguments{
\item{smc}{Surface moisture content (\%).}

\item{wind_speed}{Wind speed at 10 m (km/h).}

\item{wind_reduction}{Wind reduction factor (default 3).}
}
\value{
Fire Danger Index (m/h).
}
\description{
WA Red Book Northern Jarrah FDI
}
\references{
Beck, J.A. (1995). Equations for the forest fire behaviour
tables for Western Australia. \emph{CALMScience} 1(3):325-348.
}
