% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook.R
\name{redbook_fdi_karri}
\alias{redbook_fdi_karri}
\title{WA Red Book Karri FDI}
\usage{
redbook_fdi_karri(smc, wind_speed, wind_reduction = 3)
}
\arguments{
\item{smc}{Surface moisture content (\%).}

\item{wind_speed}{Wind speed at 10 m (km/h).}

\item{wind_reduction}{Wind reduction factor (default 3).}
}
\value{
FDI (m/h).
}
\description{
WA Red Book Karri FDI
}
