% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook.R
\name{redbook_fqcf_jarrah}
\alias{redbook_fqcf_jarrah}
\title{WA Red Book Fuel Quantity Correction Factor (Jarrah)}
\usage{
redbook_fqcf_jarrah(total_fuel, smc)
}
\arguments{
\item{total_fuel}{Total available fuel quantity (t/ha).}

\item{smc}{Surface moisture content (\%).}
}
\value{
Correction factor.
}
\description{
WA Red Book Fuel Quantity Correction Factor (Jarrah)
}
