% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook.R
\name{redbook_nwc_dry}
\alias{redbook_nwc_dry}
\title{WA Red Book Night Wetting Correction (Dry)}
\usage{
redbook_nwc_dry(smc_prev, rh_count, days_since_rain)
}
\arguments{
\item{smc_prev}{Previous 1500h surface moisture content (\%).}

\item{rh_count}{RH count value.}

\item{days_since_rain}{Days since rain.}
}
\value{
Night wetting correction for dry conditions.
}
\description{
WA Red Book Night Wetting Correction (Dry)
}
