% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook.R
\name{redbook_nwc_rain}
\alias{redbook_nwc_rain}
\title{WA Red Book Night Wetting Correction (Rain)}
\usage{
redbook_nwc_rain(smc_prev, rain_amount, days_since_rain)
}
\arguments{
\item{smc_prev}{Previous 1500h surface moisture content (\%).}

\item{rain_amount}{Rainfall amount (mm).}

\item{days_since_rain}{Days since rain.}
}
\value{
Night wetting correction for rain.
}
\description{
WA Red Book Night Wetting Correction (Rain)
}
