% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{rh_from_dewpoint}
\alias{rh_from_dewpoint}
\title{Calculate relative humidity from temperature and dew point}
\usage{
rh_from_dewpoint(temp, dew_pt)
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{dew_pt}{Dew point temperature (degrees C).}
}
\value{
Relative humidity (\%).
}
\description{
Calculate relative humidity from temperature and dew point
}
