% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{rh_from_wetbulb}
\alias{rh_from_wetbulb}
\title{Calculate RH and dew point from dry and wet bulb temperatures}
\usage{
rh_from_wetbulb(temp, wet_bulb)
}
\arguments{
\item{temp}{Dry bulb temperature (degrees C).}

\item{wet_bulb}{Wet bulb temperature (degrees C).}
}
\value{
A list with components \code{rh} (\%) and \code{dew_pt} (degrees C).
}
\description{
Uses the psychrometric equation at standard pressure (101.3 kPa).
}
