% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope_convert}
\alias{slope_convert}
\title{Convert slope between percent and degrees}
\usage{
slope_convert(value, from = c("percent", "degrees"))
}
\arguments{
\item{value}{Slope value to convert.}

\item{from}{Unit of input: "percent" or "degrees".}
}
\value{
Converted slope value.
}
\description{
Convert slope between percent and degrees
}
