% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope_factor}
\alias{slope_factor}
\title{Rate of spread slope modification factor}
\usage{
slope_factor(slope)
}
\arguments{
\item{slope}{Slope in degrees.}
}
\value{
Multiplicative factor for rate of spread.
}
\description{
Standard exponential slope correction used across multiple Australian
fire behaviour models.
}
