% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope_from_rise}
\alias{slope_from_rise}
\title{Calculate slope from rise and distance}
\usage{
slope_from_rise(rise, distance)
}
\arguments{
\item{rise}{Elevation change (m).}

\item{distance}{Horizontal distance (m).}
}
\value{
Slope in degrees.
}
\description{
Calculate slope from rise and distance
}
