% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope_solve}
\alias{slope_solve}
\title{Find rise or distance given slope and one measurement}
\usage{
slope_solve(slope_deg, rise = NULL, distance = NULL)
}
\arguments{
\item{slope_deg}{Slope (degrees).}

\item{rise}{Rise (m). Provide either rise or distance.}

\item{distance}{Distance (m). Provide either rise or distance.}
}
\value{
The missing measurement (rise or distance in m).
}
\description{
Find rise or distance given slope and one measurement
}
