% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope_template_distance}
\alias{slope_template_distance}
\title{Calculate horizontal distance for a slope template}
\usage{
slope_template_distance(rise, slope_deg)
}
\arguments{
\item{rise}{Elevation change between contours (m).}

\item{slope_deg}{Slope (degrees).}
}
\value{
Horizontal distance (m).
}
\description{
Distance on a map between contours for a given rise and slope.
}
