% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppression.R
\name{tanker_rate}
\alias{tanker_rate}
\title{Tanker construction rate}
\usage{
tanker_rate(n_tankers, access, bark_hazard, flame_height, wind_speed)
}
\arguments{
\item{n_tankers}{Number of tankers (2500-4000L).}

\item{access}{Road access quality: "Good" or "Poor".}

\item{bark_hazard}{Bark hazard: "Low", "Moderate", "High", "Very High",
or "Extreme".}

\item{flame_height}{Flame height (m).}

\item{wind_speed}{Wind speed (km/h).}
}
\value{
Estimated construction rate (m/h).
}
\description{
Tanker construction rate
}
