% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_vesta.R
\name{vesta_fuel_moisture}
\alias{vesta_fuel_moisture}
\title{VESTA fine fuel moisture content}
\usage{
vesta_fuel_moisture(temp, rh, table = c("summer_pm", "daytime", "night"))
}
\arguments{
\item{temp}{Air temperature (degrees C).}

\item{rh}{Relative humidity (\%).}

\item{table}{Which moisture table to use: "summer_pm" (Table M1, summer
13:00-17:00), "daytime" (Table M2), or "night" (Table M3).}
}
\value{
Fine fuel moisture content (\%).
}
\description{
Three models for different times of day.
}
\references{
Equations derived by Kevin Tolhurst based on VESTA field
guide (2007).
}
