% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_vesta.R
\name{vesta_moisture_table}
\alias{vesta_moisture_table}
\title{Determine which VESTA moisture table to use}
\usage{
vesta_moisture_table(month, hour)
}
\arguments{
\item{month}{Month (1-12).}

\item{hour}{Hour of day (0-23).}
}
\value{
Character string: "summer_pm", "daytime", or "night".
}
\description{
Determine which VESTA moisture table to use
}
