% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_reduction_factors}
\alias{wind_reduction_factors}
\title{Wind Reduction Factor Guide}
\format{
A data frame with 10 rows and 3 columns:
\describe{
\item{factor_low}{Lower numeric wind reduction factor}
\item{factor_high}{Upper numeric wind reduction factor}
\item{vegetation}{Vegetation type description}
}
}
\source{
Tolhurst (2011) revised wind reduction factors.
}
\usage{
wind_reduction_factors
}
\description{
Lookup table for wind reduction factors by vegetation type.
}
\keyword{datasets}
